Extracted functionality into internal/importer
